requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
httpx==0.25.1
websockets==12.0
nats-py==2.12.0